Добавляет компонент пункта маршрута
